refactor(front): File reorganisation, code cleaning
