+++ frontend/src/app/projects/models/add-t4c-source/add-t4c-source.component.spec.ts
@@ -7,7 +7,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { AddT4cSourceComponent } from './add-t4c-source.component';
 
-describe('AddT4cSourceComponent', () => {
+xdescribe('AddT4cSourceComponent', () => {
   let component: AddT4cSourceComponent;
   let fixture: ComponentFixture<AddT4cSourceComponent>;
