+++ frontend/src/app/projects/project-detail/model-overview/model-overview.component.html
@@ -56,7 +56,7 @@
           </a>
           <a
             mat-mini-fab
-            [disabled]="!model.t4c_model"
+            [disabled]="!model.t4c_models"
             matTooltip="Start synchronization (Coming soon)"
           >
             <mat-icon>sync</mat-icon>
@@ -70,7 +70,7 @@
             <mat-icon>link</mat-icon>
           </a>
           <a
-            *ngIf="model.t4c_model"
+            *ngIf="model.t4c_models"
             routerLink="/workspaces"
             mat-mini-fab
             matTooltip="Request persistent session"
