+++ frontend/src/app/services/model/model.service.ts
@@ -11,7 +11,7 @@ import {
   ToolType,
   ToolVersion,
 } from 'src/app/settings/core/tools-settings/tool.service';
-import { T4CModel } from 'src/app/services/source/t4c-model.service';
+import {T4CModel} from "src/app/services/modelsources/t4c-model/t4c-model.service";
 import { environment } from 'src/environments/environment';
 
 export interface NewModel {
@@ -29,7 +29,7 @@ export interface Model {
   tool: Tool;
   version: ToolVersion | null;
   type: ToolType | null;
-  t4c_model: T4CModel[];
+  t4c_models: T4CModel[];
   git_model_id: number | null;
 }
