+++ frontend/src/app/services/modelsources/t4c-model/t4c-model.service.ts
@@ -6,8 +6,8 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { BehaviorSubject, Observable } from 'rxjs';
-import { environment } from '../../../environments/environment';
-import { T4CRepository } from '../../settings/modelsources/t4c-settings/service/t4c-repos/t4c-repo.service';
+import { environment } from '../../../../environments/environment';
+import { T4CRepository } from '../../../settings/modelsources/t4c-settings/service/t4c-repos/t4c-repo.service';
 
 @Injectable({
   providedIn: 'root',
