+++ frontend/src/app/settings/modelsources/t4c-settings/service/t4c-repos/t4c-repo.service.ts
@@ -7,7 +7,7 @@ import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { BehaviorSubject, Observable } from 'rxjs';
 import { ToastService } from 'src/app/helpers/toast/toast.service';
-import { T4CInstance } from 'src/app/services/settings/t4c-model.service';
+import { T4CInstance } from 'src/app/services/settings/t4c-instance.service';
 import { environment } from 'src/environments/environment';
 
 @Injectable({
