+++ frontend/src/app/settings/modelsources/t4c-settings/t4c-instance-settings/t4c-instance-settings.component.ts
@@ -21,7 +21,7 @@ import {
 import { MatDialog } from '@angular/material/dialog';
 import { MatSelectionList } from '@angular/material/list';
 import { tap, switchMap } from 'rxjs';
-import { T4CInstance } from 'src/app/services/settings/t4c-model.service';
+import { T4CInstance } from 'src/app/services/settings/t4c-instance.service';
 import {
   CreateT4CRepository,
   T4CRepoService,
