+++ frontend/src/app/settings/modelsources/t4c-settings/t4c-settings.component.ts
@@ -9,7 +9,7 @@ import { NavBarService } from 'src/app/general/navbar/service/nav-bar.service';
 import {
   T4CInstance,
   T4CInstanceService,
-} from '../../../services/settings/t4c-model.service';
+} from '../../../services/settings/t4c-instance.service';
 
 @Component({
   selector: 'app-t4c-settings',
